requirements.txt
gspread
oauth2client
playwright
python-dotenv
aiohttp